Corrected problems and  made undo and redo work
Problems still with rename and make dir in undo and redo because in order to undo-redo them you need the new name of the directories or files. Somehow I get problems with that. I have to work on that and on recycle next. Otherwise it works fine and as far as I can tell it can be used with out problems. In all the tests I had no problems. Still since it is a program that changes the file system I advice caution. It is best if it is used in a Virtual Machine.