multiple paste problem
I corrected the problem of multiple paste that would result in data loss. Since I did not want to use big parts of memory to store big files, I prefered to make paste possible only once. If you wish to paste it again you need to select it and "copy" or "cut" it before you push the "paste" button again