Copy and move work ok with overwrite control
In this commit I finished with move overwrite control. The undo and redo work ok but the do not conform with Microsoft Windows Explorer, so if use use undo with the overwrite control you might lose data. I will fix this next. I will make undo to have similar behaviour with Microsoft Windows Explorer so that there is no data loss because of unexpected behaviour.