Rewrote "delete" and fileaction subroutines, and implemented undo and redo with overwrite control
In this commit I rewrote the delete command. Now it works better. Have in mind that delete is permanent 
and cannot be undone so be careful because what you delete is deleted permanently and cannot be undone.

In this commit I also rewrote the overwrite control for copy and cut (move) It is now implemented 
and tested. After many tests it seems that it works ok. Undo and redo work ok  with overwrite 
control and it is similar with the linux file manager. That means that it keeps the files that you chose
not to overwrite at target when you do undo and the original files you did not copy or move (chose to 
keep the target files at the target -so the source files reamained at source) when you do redo.

I also rewrote some parts of code (filaction subroutine etc. to make it more readable)